level2/window.h,cpp: move SolidBackground to a dedicated header so it's available outside Window
